+++ projects/ngx-dragdrop/src/public_api.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of ngx-dragdrop
+ */
+
+export * from './lib/dragDrop.service';
+export * from './lib/draggable.directive';
+export * from './lib/droppable.directive';
+export * from './lib/ngx-dragdrop.module';