+++ src/index.ts
@@ -1,2 +1,0 @@
-/* Copyright VMware, Inc. All rights reserved. VMware Confidential */
-export {NgxDragDropModule} from "./ngx-dragdrop.module";