+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 6.0.0-beta.1
+### Features
+* Updated Angular to 6.0.0
+* Added `IOnDropData` interface for OnDrop $event type
+### Breaking changes
+* `ngx-draggable` changed to `ngxDraggable`
+* `ngx-dragTag` changed to `ngxDragTag`
+* `ngx-draggedClass` changed to `ngxDraggedClass`
+* `ngx-dragData` changed to `ngxDragData`
+* `ngx-droppable` changed to `ngxDroppable`
+* `ngx-dropData` changed to `ngxDropData`
+* `ngx-dropTags` changed to `ngxDropTags`
+
 ## 1.0.0
 ### Features
 * Updated Angular to 4.0.0